separate CameraModel for filters
